Remove support for non-object plugins to be added.
Reviewed all instances of `registry.add` in emberobserver.com, and
found only a single instance (out of > 75 addons using `registry.add`)
that was utilizing `registry.add(<type>, <string>);` (and that addon
hasn't been updated/touched in > 2 years).

Removing this dead code, and simplifying the rest of the responsibilities
of this repo.